+++ src/game/mod.rs
@@ -14,6 +14,7 @@ pub(crate) mod components;
 pub(crate) mod constants;
 pub(crate) mod crt_effect;
 pub(crate) mod drops;
+pub(crate) mod game_mode;
 pub(crate) mod input;
 mod loading;
 pub(crate) mod messages;